Make old sample visible as  third view mode on Sample List
remove duplicated code
